+++ tests/testthat/test-Rogues.R
@@ -1,0 +1,7 @@
+test_that("Rogues found with TipInstability()", {
+  
+  library("TreeTools", quietly = TRUE)
+  trees <- AddTipEverywhere(BalancedTree(8), 'Rogue')
+  instab <- TipInstability(trees)
+  expect_equal('Rogue', names(which.max(instab)))
+})