+++ NEWS.md
@@ -1,10 +1,11 @@
-# TreeSearch 0.4.3.9008 (development)
+# TreeSearch 0.4.3.9009 (development)
 
  * v......9007 introduces a pre-ratchet TBR search and adjusts default settings.
  * v0.4.3.9005 fixes some bugs that may affect tree scoring under IW & PP.
 
  - `EasyTrees()` 'shiny' graphical user interface for tree search
  - `PlotCharacter()` reconstructs character distributions on trees.
+ - `TipInstability()` finds rogue taxa.
  - `SiteConcordance()` calculates exact site concordance
  - `ConcordantInformation()` evaluates signal:noise of dataset implied by a
    given tree.